Extracts FeedStoreSpecs checks into reusable assertions.
