--- FeedModule/FeedModuleTests/Feed Cache/XCTestCase+FeedStoreSpecs.swift
@@ -1,67 +1,0 @@
-//
-//  XCTestCase+FeedStoreSpecs.swift
-//  FeedModuleTests
-//
-//  Created by Omran Khoja on 2/10/22.
-//
-
-import XCTest
-import FeedModule
-
-extension FeedStoreSpecs where Self: XCTestCase {
-    
-    @discardableResult
-    func insert(_ cache: (feed: [LocalFeedImage], timestamp: Date), to sut: FeedStore, file: StaticString = #filePath, line: UInt = #line) -> Error? {
-        var insertionError: Error?
-        let exp = expectation(description: "Wait for insertion")
-        sut.insert(cache.feed, timestamp: cache.timestamp) { receivedError in
-            insertionError = receivedError
-            exp.fulfill()
-        }
-        wait(for: [exp], timeout: 1.0)
-        return insertionError
-    }
-    
-    @discardableResult
-    func deleteCache(from sut: FeedStore, file: StaticString = #filePath, line: UInt = #line) -> Error? {
-        var deletionError: Error?
-        let exp = expectation(description: "Wait for deletion")
-        sut.deleteCachedFeed() { receivedError in
-            deletionError = receivedError
-            exp.fulfill()
-        }
-        wait(for: [exp], timeout: 20.0)
-        return deletionError
-    }
-    
-    func expect(_ sut: FeedStore, toRetrieveTwice expectedResult: RetrieveCachedFeedResult, file: StaticString = #filePath, line: UInt = #line) {
-        expect(sut, toRetrieve: expectedResult, file: file, line: line)
-        expect(sut, toRetrieve: expectedResult, file: file, line: line)
-    }
-
-    func expect(_ sut: FeedStore, toRetrieve expectedResult: RetrieveCachedFeedResult, file: StaticString = #filePath, line: UInt = #line) {
-        let exp = expectation(description: "Wait for result")
-        sut.retrieve { retrievedResult in
-                switch (retrievedResult, expectedResult) {
-                case (.empty, .empty), (.failure, .failure):
-                    break;
-                case let (.found(expectedFeed, expectedTimestamp), .found(retrievedFeed, retrievedTimestamp)):
-                    XCTAssertEqual(expectedFeed, retrievedFeed, file: file, line: line)
-                    XCTAssertEqual(expectedTimestamp, retrievedTimestamp, file: file, line: line)
-                default:
-                    XCTFail("Exepected to retrieve \(expectedResult), got \(retrievedResult) instead", file: file, line: line)
-                }
-            exp.fulfill()
-        }
-            
-        wait(for: [exp], timeout: 1.0)
-    }
-    
-    func testSpecificStoreURL() -> URL {
-        return cachesDirectory().appendingPathComponent("\(type(of: self)).store")
-    }
-    
-    func cachesDirectory() -> URL {
-        return FileManager.default.urls(for: .cachesDirectory, in: .userDomainMask).first!
-    }
-}
